+++ docs/source/changelog.md
@@ -4,6 +4,7 @@
 - Fix numpy import issue on OS X
 - CI: Use SQL Server docker image instead of Azure Edge
 - Improve documentation for materialization details.
+- Add support for columnstore tables in `MSSQLTableStore` via `MSSqlMaterializationDetails`
 
 ## 0.9.10 (2025-03-18)
 - Fix incompatibility with pydiverse.transform >=0.2.1 (<0.2.0 still supported) (0.2.0 will not be supported):